change the data type from bigInteger tio integer

+++ app/Models/CategoryJoin.php
@@ -14,9 +14,9 @@ class CategoryJoin extends Model
     public $table="category_joins" ;
 
     protected $casts = [
-        'id' => 'bigInteger',   
-        'post_id' => 'bigInteger',   
-        'category_id' => 'bigInteger',   
+        'id' => 'integer',   
+        'post_id' => 'integer',   
+        'category_id' => 'integer',   
     ];
 
     protected $fillable = [
